requirements.txt
colorama
KahootPY
keyboard